SC-3788 rename Messenger to meet naming convention
