+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 - SC-7447 Add warning text for links when leaving the schul-cloud platform
 - SC-6060 Updated caniuse-lite
+- SC-3788 Renamed Messenger component to meet naming convention
 
 ## 25.2.0
